memos.jsonが存在しない場合に起動できない不具合を修正
存在しないファイルを読み込もうとして起動できなかったので修正しました。

+++ memo.rb
@@ -69,6 +69,8 @@ class Memo
   end
 
   def self.load
+    return unless File.exist?(@@JSON_FILE)
+
     memos_json = {}
     File.open(@@JSON_FILE, 'r') { |file| memos_json = JSON.parse(file.readline, symbolize_names: true)}
